refactor(layout): move Layout directory out of components
